+++ scalr/data_source_scalr_access_policy_test.go
@@ -27,7 +27,7 @@ func TestAccScalrAccessPolicyDataSource_basic(t *testing.T) {
 			},
 			{
 				Config:      testAccAccessPolicyDataSourceNotFoundConfig(),
-				ExpectError: regexp.MustCompile("IamAccessPolicy with ID 'ap-123' not found or user unauthorized"),
+				ExpectError: regexp.MustCompile("AccessPolicy 'ap-123' not found"),
 				PlanOnly:    true,
 			},
 		},